refactor: move schema-related classes to src/parse
